Separate media serializers based on usage
