+++ api/catalog/api/views/audio_views.py
@@ -15,13 +15,15 @@ from catalog.api.docs.audio_docs import (
     AudioThumbnail,
 )
 from catalog.api.models import Audio
-from catalog.api.serializers.audio_serializers import (
+from catalog.api.serializers.request.audio import (
     AudioReportSerializer,
     AudioSearchRequestSerializer,
+)
+from catalog.api.serializers.request.media import MediaThumbnailRequestSerializer
+from catalog.api.serializers.response.audio import (
     AudioSerializer,
     AudioWaveformSerializer,
 )
-from catalog.api.serializers.request.media import MediaThumbnailRequestSerializer
 from catalog.api.utils.exceptions import get_api_exception
 from catalog.api.utils.throttle import OneThousandPerMinute
 from catalog.api.views.media_views import MediaViewSet
